requirements.txt
aniso8601==9.0.1
blinker==1.8.1
click==8.1.7
dnspython==2.6.1
Flask==3.0.3
Flask-JWT-Extended==4.6.0
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.10
flask-restful-swagger
flask-swagger-ui==4.11.1
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
pydantic
pydantic[email]
numpy==1.26.4
pandas==2.2.2
PyJWT==2.8.0
pymongo==4.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
tzdata==2024.1
Werkzeug==3.0.2